refactor: split CCtypeAsciiScope into its own header
The helper now has a second use case (locale-safe lowercasing of
ISO country codes in the libmaxminddb path), so the file it lived
in is no longer the right home. Move CCtypeAsciiScope verbatim
into src/CCtypeAsciiScope.h; have CamuleFileConfig.h include it.

No behaviour change.
